requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
selenium
pandas
numpy
groq
python-dotenv
sqlite3
